refactor: extract neutral command support
