+++ crates/tui/src/tools/subagent/mod.rs
@@ -5140,7 +5140,7 @@ fn fallback_subagent_assignment_route(
     let model = if let Some(model) = configured_model {
         model
     } else if runtime.auto_model {
-        crate::commands::auto_model_heuristic(prompt, &runtime.model)
+        crate::model_routing::auto_model_heuristic(prompt, &runtime.model)
     } else {
         runtime.model.clone()
     };
@@ -5166,7 +5166,7 @@ fn fallback_subagent_assignment_route(
 async fn subagent_flash_router(
     runtime: &SubAgentRuntime,
     prompt: &str,
-) -> Result<Option<crate::commands::AutoRouteRecommendation>> {
+) -> Result<Option<crate::model_routing::AutoRouteRecommendation>> {
     if cfg!(test) {
         return Ok(None);
     }
@@ -5199,7 +5199,7 @@ async fn subagent_flash_router(
         runtime.client.create_message(request),
     )
     .await??;
-    Ok(crate::commands::parse_auto_route_recommendation(
+    Ok(crate::model_routing::parse_auto_route_recommendation(
         &message_response_text(&response.content),
     ))
 }
