+++ crates/tui/src/tui/auto_router.rs
@@ -4,12 +4,12 @@
 //! The TUI calls `resolve_auto_model_selection` once per user turn when
 //! `app.auto_model` is set. The async function builds a recent-context
 //! summary from `api_messages` (capped to six rows of up to 900 chars
-//! each), passes it through `commands::resolve_auto_route_with_flash`,
+//! each), passes it through `model_routing::resolve_auto_route_with_flash`,
 //! and returns the selection (model + reasoning effort). The remaining
 //! helpers are pure transforms used to build that summary.
 
-use crate::commands;
 use crate::config::Config;
+use crate::model_routing;
 use crate::models::{ContentBlock, Message};
 use crate::tui::app::{App, QueuedMessage, ReasoningEffort};
 
@@ -25,13 +25,13 @@ pub(super) async fn resolve_auto_model_selection(
     config: &Config,
     message: &QueuedMessage,
     latest_content: &str,
-) -> commands::AutoRouteSelection {
+) -> model_routing::AutoRouteSelection {
     let latest_request = if latest_content.trim().is_empty() {
         message.display.as_str()
     } else {
         latest_content
     };
-    commands::resolve_auto_route_with_flash(
+    model_routing::resolve_auto_route_with_flash(
         config,
         latest_request,
         &recent_auto_router_context(&app.api_messages),
@@ -43,7 +43,7 @@ pub(super) async fn resolve_auto_model_selection(
 
 /// Normalize the heuristic effort to the canonical auto-route effort.
 pub(super) fn normalize_auto_routed_effort(effort: ReasoningEffort) -> ReasoningEffort {
-    commands::normalize_auto_route_effort(effort)
+    model_routing::normalize_auto_route_effort(effort)
 }
 
 /// Build a compact recent-context summary for the auto-route prompt.
